+++ vassal.h
@@ -1,6 +1,6 @@
 #ifndef VASSAL_H
 #define VASSAL_H
-
+#include <QtGlobal>
 
 class Vassal
 {
@@ -8,31 +8,31 @@ public:
     Vassal();
     Vassal(int, int, int, int, int);   // установка статики
 
-    void setQuantity(int);
+    void setQuantity(qint64);
     int getAttack();
     int getDefense();
     int getIncome();
 
     int getSum_attack();
     int getSum_defense();
-    int getSum_income();
+    qint64 getSum_income();
 
 private:
     // статичные
     int attack;       // атака вассала
     int defense;      // защита вассала
-    int income;       // доход от вассала
+    qint64 income;       // доход от вассала
     int st_price;     // стартовая цена
     int ch_price;     // изменение цены
 
     // динамичные
 
-    int quantity;     // колличество вассалов
+    qint64 quantity;     // колличество вассалов
     int cur_price;    // текущая цена
 
     int sum_attack;   // сумарная атака
     int sum_defense;  // сумарная защита
-    int sum_income;   // сумарный доход
+    qint64 sum_income;   // сумарный доход
 
 };
